my_currency/requirements.txt
requests==2.27.1
Django==4.0.1
djangorestframework==3.13.1
channels==4.0.0
daphne==4.0.0
channels-redis==4.1.0
python-dotenv==1.0.0
